refactor: fix package names and code (#3752)
