--- src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/player/autoQueue/ModuleAutoQueue.kt
@@ -1,33 +1,0 @@
-/*
- * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
- *
- * Copyright (c) 2015 - 2024 CCBlueX
- *
- * LiquidBounce is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * LiquidBounce is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
- */
-package net.ccbluex.liquidbounce.features.module.modules.player.autoQueue
-
-import net.ccbluex.liquidbounce.features.module.Category
-import net.ccbluex.liquidbounce.features.module.Module
-import net.ccbluex.liquidbounce.features.module.modules.player.autoQueue.modes.AutoQueueGommeDuels
-import net.ccbluex.liquidbounce.features.module.modules.player.autoQueue.modes.AutoQueueHypixelSW
-import net.ccbluex.liquidbounce.features.module.modules.player.autoQueue.modes.AutoQueuePaper
-
-object ModuleAutoQueue : Module("AutoQueue", Category.PLAYER) {
-    val modes = choices("Mode", AutoQueuePaper, arrayOf(
-        AutoQueuePaper,
-        AutoQueueHypixelSW,
-        AutoQueueGommeDuels
-    ))
-}
