--- src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/world/autoFarm/AutoFarmBlockHandler.kt
@@ -1,74 +1,0 @@
-/*
- * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
- *
- * Copyright (c) 2015 - 2024 CCBlueX
- *
- * LiquidBounce is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * LiquidBounce is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
- */
-package net.ccbluex.liquidbounce.features.module.modules.world.autoFarm
-
-import net.ccbluex.liquidbounce.utils.block.AbstractBlockLocationTracker
-import net.ccbluex.liquidbounce.utils.block.getState
-import net.minecraft.block.BlockState
-import net.minecraft.block.FarmlandBlock
-import net.minecraft.block.SoulSandBlock
-import net.minecraft.util.math.BlockPos
-
-enum class AutoFarmTrackedStates {
-    Destroy,
-    Farmland,
-    Soulsand
-}
-
-object AutoFarmBlockTracker : AbstractBlockLocationTracker<AutoFarmTrackedStates>() {
-    override fun getStateFor(pos: BlockPos, state: BlockState): AutoFarmTrackedStates? {
-        val block = state.block
-
-        if (ModuleAutoFarm.isTargeted(state, pos))
-            return AutoFarmTrackedStates.Destroy
-
-        val stateBellow = pos.down().getState() ?: return null
-
-        if (stateBellow.isAir) return null
-
-        val blockBellow = stateBellow.block
-
-        if (blockBellow is FarmlandBlock) {
-            handlePlaceableBlock(pos, state, AutoFarmTrackedStates.Farmland)
-        } else if (blockBellow is SoulSandBlock) {
-            handlePlaceableBlock(pos, state, AutoFarmTrackedStates.Soulsand)
-        }
-        if (ModuleAutoFarm.hasAirAbove(pos)) {
-            return when (block) {
-                is FarmlandBlock -> AutoFarmTrackedStates.Farmland
-                is SoulSandBlock -> AutoFarmTrackedStates.Soulsand
-                else -> null
-            }
-        }
-        return null
-    }
-
-
-    private fun handlePlaceableBlock(pos: BlockPos, state: BlockState, trackedState: AutoFarmTrackedStates) {
-        val targetBlockPos = TargetBlockPos(pos.down())
-        if (state.isAir){
-            // If there is no air above, add it
-            this.trackedBlockMap[targetBlockPos] = trackedState
-        } else {
-            // If there is no air above, we want to remove it
-            this.trackedBlockMap.remove(targetBlockPos)
-        }
-    }
-}
-
